Searching for vowels
using for loop, list and if condition to take number of names. stored the names in a list called "names". The checked if there is any vowel  in each name or not. if is there then replaced it with '*'. then asked the user if he wants to see the undated list with vowel number in details. Ans again if the user isn't interested then showed him/her total no. of vowel through put his/her whole input.